moved self-signed-cert generation to be part of role

+++ roles/hashicorp_vault_cluster/files/generate_self_signed_certs.sh
@@ -26,7 +26,7 @@ fi
 mkdir -p "$VAULT_CERTS_BASE_DIR/$VAULT_CLUSTER_FQDN"
 cd "$VAULT_CERTS_BASE_DIR/$VAULT_CLUSTER_FQDN"
 
-# If the CA cert for the FQDN doesn't exist, create it
+# If the CA cert for VAULT_CLUSTER_FQDN doesn't exist, create it
 if ! [ -f "$VAULT_CLUSTER_FQDN.ca.cert.pem" ]; then
   openssl ecparam -genkey -name prime256v1 -noout -out "$VAULT_CLUSTER_FQDN.ca.private.key"
   openssl pkcs8 -topk8 -nocrypt -in "$VAULT_CLUSTER_FQDN.ca.private.key" -out "$VAULT_CLUSTER_FQDN.ca.ec.private.key"
@@ -38,15 +38,17 @@ if ! [ -f "$VAULT_CLUSTER_FQDN.ca.cert.pem" ]; then
     -out "$VAULT_CLUSTER_FQDN.ca.cert.pem"
 fi
 
-mkdir -p "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN"
-cd "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN"
+# If the VAULT_CLUSTER_MEMBER cert doesn't exist, create it
+if ! [ -f "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN/$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.cert.pem" ]; then
+  mkdir -p "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN"
+  cd "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN"
 
-# create new private key
-openssl ecparam -genkey -name prime256v1 -noout -out "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.ec.private.key"
-openssl pkcs8 -topk8 -nocrypt -in "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.ec.private.key" -out "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.private.key"
+  # create new private key
+  openssl ecparam -genkey -name prime256v1 -noout -out "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.ec.private.key"
+  openssl pkcs8 -topk8 -nocrypt -in "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.ec.private.key" -out "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.private.key"
 
-# configuartion file for domain
-cat <<EOF>"$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.openssl.cfg"
+  # configuartion file for domain
+  cat <<EOF>"$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.openssl.cfg"
 [req]
 distinguished_name = req_distinguished_name
 req_extensions = v3_req
@@ -65,19 +67,21 @@ subjectAltName = @alt_names
 DNS.1 = $VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN
 EOF
 
-# generate a certificate sign request with the config
-openssl req -new \
-  -key "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.private.key" -config "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.openssl.cfg" \
-  -out "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.csr"
+  # generate a certificate sign request with the config
+  openssl req -new \
+    -key "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.private.key" -config "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.openssl.cfg" \
+    -out "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.csr"
 
-openssl req -x509 \
-  -sha256 \
-  -days 365 \
-  -in "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.csr" \
-  -CA "../$VAULT_CLUSTER_FQDN.ca.cert.pem" \
-  -CAkey "../$VAULT_CLUSTER_FQDN.ca.ec.private.key" \
-  -copy_extensions copyall \
-  -addext "basicConstraints = critical,CA:FALSE" \
-  -out "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.cert.pem"
+  openssl req -x509 \
+    -sha256 \
+    -days 365 \
+    -in "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.csr" \
+    -CA "../$VAULT_CLUSTER_FQDN.ca.cert.pem" \
+    -CAkey "../$VAULT_CLUSTER_FQDN.ca.ec.private.key" \
+    -copy_extensions copyall \
+    -addext "basicConstraints = critical,CA:FALSE" \
+    -out "$VAULT_CLUSTER_MEMBER.$VAULT_CLUSTER_FQDN.cert.pem"
 
-cp "../$VAULT_CLUSTER_FQDN.ca.cert.pem" .
+  cp "../$VAULT_CLUSTER_FQDN.ca.cert.pem" .
+
+fi
